requirements_enhanced.txt
# YouTube Transcript Summarizer - Enhanced with TextRank & PCA
# Requirements file

# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# YouTube transcript extraction
youtube-transcript-api>=0.6.1

# Google Gemini AI
google-generativeai>=0.3.0

# Database (PostgreSQL - Neon)
psycopg2-binary>=2.9.9

# NLP and Machine Learning
nltk>=3.8.1
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Graph algorithms for TextRank
networkx>=3.1

# Additional utilities
requests>=2.31.0
pandas>=2.0.0
